Refactored testing environment.
